fix(actions): extract sync helpers from 'use server' modules
Next.js production build fails with "Server Actions must be async
functions" when sync helpers are exported from 'use server' files.
Move maskPhone, interpolateTemplate (sms) and addBusinessDays
(quotes) into src/lib/sms/helpers.ts and src/lib/quotes/helpers.ts
respectively. Server actions still consume them via internal import.
