+++ src/actions/admin/quotes.ts
@@ -2,6 +2,7 @@
 
 import { revalidatePath } from 'next/cache';
 import { getStaffCtx } from '@/lib/auth';
+import { addBusinessDays } from '@/lib/quotes/helpers';
 
 export type QuoteLineItem = {
   description: string;
@@ -33,21 +34,6 @@ export type Quote = {
   created: string;
   updated: string;
 };
-
-/**
- * Add N business days (Mon–Fri) to a date.
- * RD 1457/1986 requires quote validity of at least 12 business days.
- */
-export function addBusinessDays(date: Date, days: number): Date {
-  const result = new Date(date);
-  let added = 0;
-  while (added < days) {
-    result.setDate(result.getDate() + 1);
-    const dow = result.getDay();
-    if (dow !== 0 && dow !== 6) added++;
-  }
-  return result;
-}
 
 export type GetQuotesResult =
   | { ok: true; quotes: Quote[] }
