+++ src/lib/quotes/helpers.ts
@@ -1,0 +1,14 @@
+/**
+ * Add N business days (Mon–Fri) to a date.
+ * RD 1457/1986 requires quote validity of at least 12 business days.
+ */
+export function addBusinessDays(date: Date, days: number): Date {
+  const result = new Date(date);
+  let added = 0;
+  while (added < days) {
+    result.setDate(result.getDate() + 1);
+    const dow = result.getDay();
+    if (dow !== 0 && dow !== 6) added++;
+  }
+  return result;
+}
